w2parser/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.8.1
attrs==23.2.0
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.5
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
frozenlist==1.4.1
g4f==0.3.0.7
greenlet==3.0.3
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.46
langchain-experimental==0.0.57
langchain-text-splitters==0.0.1
langsmith==0.1.52
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.1
packaging==23.2
pandas==2.2.2
pdf2image==1.17.0
pillow==10.3.0
poppler-utils==0.1.0
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.29
sqlparse==0.5.0
tenacity==8.2.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
